Issue #29: Add conda-based packaging script

+++ packaging/build_appimage.sh
@@ -29,14 +29,153 @@ pushd /tmp
 if [ -e appimagetool-x86_64.AppImage ]; then
     rm appimagetool-x86_64.AppImage
 fi
-wget https://github.com/probonopd/AppImageKit/releases/download/12/appimagetool-x86_64.AppImage
+#wget https://github.com/probonopd/AppImageKit/releases/download/12/appimagetool-x86_64.AppImage
+wget http://archive.main.int/archive/appimagetool-x86_64.AppImage
 chmod +x appimagetool-x86_64.AppImage
 popd
 popd
 
-DEBOOTSTRAP=1
+DEBOOTSTRAP=0
+MINICONDA=1
 
-if [ ${DEBOOTSTRAP} -eq 1 ]; then
+MINICONDA_PKG=Miniconda3-py38_4.8.3-Linux-x86_64.sh
+
+if [ ${MINICONDA} -eq 1 ]; then
+    echo "Installing miniconda"
+    pushd ${APPDIR}
+    PKG=${MINICONDA_PKG}
+    if [ ! -e ${PKG} ]; then
+        wget http://archive.main.int/archive/${PKG}
+        bash ${PKG} -b -p usr -f
+    fi
+    # appdir_python()
+    # {
+    #     env LD_LIBRARY_PATH="${APPDIR}/usr/lib" "${APPDIR}/usr/bin/python3.8" -s "$@"
+    # }
+    # python='appdir_python'
+    echo "Activating envinronment"
+    source usr/bin/activate
+    echo "Installing deps"
+    usr/bin/pip install --force-reinstall PyQt5 PyOpenGL PyOpenGL_accelerate numpy watchdog pyinotify six ply
+    echo "pip: $(which -a pip)"
+
+    # if [ ! -e ply ]; then
+    #     git clone https://github.com/dabeaz/ply.git
+    #     pushd ply
+    #     git checkout e5d40872956764a47dbf9df6a455568f61f92173 -b build
+    #     cp -pr ply ${APPDIR}/usr/lib/python3.8/site-packages
+    #     popd
+    # fi
+
+    if [ ! -e ezdxf ]; then
+        git clone git@github.com:snegovick/ezdxf.git
+        pushd ezdxf
+        git checkout 1070c67779f75c707c8817b2cc2eca87154fdab5 -b build
+        ${APPDIR}/usr/bin/python3.8 setup.py build -j$(nproc) install --prefix ${APPDIR}/usr
+        popd
+        #mv ${APPDIR}/usr/lib/python3.8/site-packages/ezdxf-0.11b1-py3.8.egg/ezdxf ${APPDIR}/usr/lib/python3.8
+        #mv ${APPDIR}/usr/lib/python3.8/site-packages/pyparsing-3.0.0a2-py3.8.egg/pyparsing ${APPDIR}/usr/lib/python3.8
+    fi
+
+    popd
+
+    echo "Copy all modules into python3.8 path"
+    # rm -rf ${APPDIR}/usr/lib/python3/dist-packages/__pycache__
+    # mv ${APPDIR}/usr/lib/python3/dist-packages/* ${APPDIR}/usr/lib/python3.8
+    mv ${APPDIR}/usr/lib/python3/site-packages/* ${APPDIR}/usr/lib/python3.8
+
+    ${APPDIR}/usr/bin/python3.8 setup.py build -j$(nproc) install --prefix ${APPDIR}/usr
+    pushd ${APPDIR}/usr/lib/python3.8/site-packages
+    # unzip bcad-*.egg
+    # mv bcad ${APPDIR}/usr/lib/python3.8
+    popd
+
+    pushd ${APPDIR}
+
+    echo "Clean up image"
+    
+    # rm -f bin
+    # rm -rf boot
+    # rm -rf etc
+    # rm -f dev
+    rm -rf ezdxf
+    # rm -f lib32
+    # #rm -f lib64
+    # rm -rf media
+    # rm -rf mnt
+    rm -rf occ_build
+    rm -rf opencascade-7.4.0
+    rm opencascade-7.4.0.tgz
+    # rm -rf opt
+    # rm -f proc
+    rm -rf pyocc_build
+    rm -rf pythonocc-core
+    rm  ${MINICONDA_PKG}
+    # rm -rf run
+    # rm -f sbin
+    # rm -rf srv
+    # rm -rf sys
+    # rm -rf tmp
+    # rm -rf var
+    # rm -rf home
+    # rm -f libx32
+
+    pushd usr/bin
+    find . ! -name python3.8 ! -name bcad-launcher -maxdepth 1 -type f -delete
+    popd
+
+    ACTUALLY_RM=0
+
+    # pushd usr/lib/x86_64-linux-gnu
+    # if [ -e ${ROOTDIR}/packaging/keep.list ]; then
+    #     for i in *; do
+    #         if ! grep -qxFe "$i" ${ROOTDIR}/packaging/keep.list; then
+    #             if [ ${ACTUALLY_RM} -eq 1 ]; then
+    #                 echo "Deleting: $i"
+    #                 rm -rf "$i"
+    #             else
+    #                 echo "Pretending to delete $i"
+    #             fi
+    #         fi
+    #     done
+    # else
+    #     echo "Error: ${ROOTDIR}/packaging/keep.list is missing"
+    #     exit 1
+    # fi
+    # popd
+
+    ACTUALLY_RM=1
+    pushd usr/lib
+    if [ -e ${ROOTDIR}/packaging/keep.list ]; then
+        for i in *; do
+            if ! grep -qxFe "$i" ${ROOTDIR}/packaging/keep.list; then
+                if [ ${ACTUALLY_RM} -eq 1 ]; then
+                    echo "Deleting: $i"
+                    rm -rf "$i"
+                else
+                    echo "Pretending to delete $i"
+                fi
+            fi
+        done
+    else
+        echo "Error: ${ROOTDIR}/packaging/keep.list is missing"
+        exit 1
+    fi
+    popd
+
+
+    rm -rf usr/compiler_compat
+    rm -rf usr/conda-meta
+    rm -rf usr/condabin
+    rm -rf usr/envs
+    rm -rf usr/etc
+    rm -rf usr/include
+    rm -rf usr/pkgs
+    rm -rf usr/share
+    rm -rf usr/shell
+    rm -rf usr/ssl
+    mv usr/LICENSE.txt usr/CONDA_LICENSE.txt
+elif [ ${DEBOOTSTRAP} -eq 1 ]; then
     echo "Running dbst"
     bash packaging/run-dbst.sh bcad.AppDir
     pushd ${APPDIR}
@@ -64,6 +203,10 @@ if [ ${DEBOOTSTRAP} -eq 1 ]; then
     mv ${APPDIR}/usr/lib/python3/site-packages/* ${APPDIR}/usr/lib/python3.8
 
     ${APPDIR}/usr/bin/python3.8 setup.py build -j$(nproc) install --prefix ${APPDIR}/usr
+    pushd ${APPDIR}/usr/lib/python3.8/site-packages
+    unzip bcad-*.egg
+    mv bcad ${APPDIR}/usr/lib/python3.8
+    popd
     
     pushd ${APPDIR}
 
@@ -74,9 +217,8 @@ if [ ${DEBOOTSTRAP} -eq 1 ]; then
     rm -rf etc
     rm -f dev
     rm -rf ezdxf
-    rm -f lib
     rm -f lib32
-    rm -f lib64
+    #rm -f lib64
     rm -rf media
     rm -rf mnt
     rm -rf occ_build
@@ -93,18 +235,25 @@ if [ ${DEBOOTSTRAP} -eq 1 ]; then
     rm -rf sys
     rm -rf tmp
     rm -rf var
+    rm -rf home
+    rm -f libx32
 
     pushd usr/bin
     find . ! -name python3.8 ! -name bcad-launcher -maxdepth 1 -type f -delete
     popd
 
     pushd usr/lib/x86_64-linux-gnu
+    ACTUALLY_RM=0
 
     if [ -e ${ROOTDIR}/packaging/keep.list ]; then
         for i in *; do
             if ! grep -qxFe "$i" ${ROOTDIR}/packaging/keep.list; then
-                echo "Deleting: $i"
-                rm -rf "$i"
+                if [ ${ACTUALLY_RM} -eq 1 ]; then
+                    echo "Deleting: $i"
+                    rm -rf "$i"
+                else
+                    echo "Pretending to delete $i"
+                fi
             fi
         done
     else
@@ -117,8 +266,12 @@ if [ ${DEBOOTSTRAP} -eq 1 ]; then
     if [ -e ${ROOTDIR}/packaging/keep.list ]; then
         for i in *; do
             if ! grep -qxFe "$i" ${ROOTDIR}/packaging/keep.list; then
-                echo "Deleting: $i"
-                rm -rf "$i"
+                if [ ${ACTUALLY_RM} -eq 1 ]; then
+                    echo "Deleting: $i"
+                    rm -rf "$i"
+                else
+                    echo "Pretending to delete $i"
+                fi
             fi
         done
     else
@@ -281,7 +434,7 @@ cat >"AppRun" <<\EOF
 #!/bin/sh
 set -e
 APPDIR="$(dirname "$(readlink -e "$0")")"
-export LD_LIBRARY_PATH="${APPDIR}/usr/lib/:${APPDIR}/usr/x86_64-linux-gnu${LD_LIBRARY_PATH+:$LD_LIBRARY_PATH}"
+export LD_LIBRARY_PATH="${APPDIR}/usr/lib/:${APPDIR}/lib/x86_64-linux-gnu:${APPDIR}/usr/lib/x86_64-linux-gnu:${LD_LIBRARY_PATH}"
 export PATH="${APPDIR}/usr/bin:${PATH}"
 exec "${APPDIR}/usr/bin/python3.8" -m bcad.binterpreter.scl "$@"
 EOF
